# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Predixa AI Enterprise","description":"Enterprise AI Lottery Prediction Platform","version":"2.3.0"},"paths":{"/":{"get":{"tags":["Root"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Swagger OAuth2 login.\nUsername = email\nPassword = password","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["Authentication"],"summary":"Me","operationId":"me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/dashboard":{"get":{"tags":["Dashboard"],"summary":"Get Dashboard","description":"Enterprise Dashboard\n\nAccess:\n    Authenticated users only.","operationId":"get_dashboard_dashboard_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/health":{"get":{"summary":"Health Check","description":"Check if the application is healthy.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/version":{"get":{"summary":"Get Version","description":"Get the version of the application.","operationId":"get_version_version_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/draws/":{"get":{"tags":["Draws"],"summary":"Get Draws","operationId":"get_draws_draws__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DrawResponse"},"type":"array","title":"Response Get Draws Draws  Get"}}}}}},"post":{"tags":["Draws"],"summary":"Create Draw","operationId":"create_draw_draws__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DrawCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DrawResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/draws/{draw_id}":{"get":{"tags":["Draws"],"summary":"Get Draw","operationId":"get_draw_draws__draw_id__get","parameters":[{"name":"draw_id","in":"path","required":true,"schema":{"type":"integer","title":"Draw Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DrawResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Draws"],"summary":"Update Draw","operationId":"update_draw_draws__draw_id__put","parameters":[{"name":"draw_id","in":"path","required":true,"schema":{"type":"integer","title":"Draw Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DrawUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DrawResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Draws"],"summary":"Delete Draw","operationId":"delete_draw_draws__draw_id__delete","parameters":[{"name":"draw_id","in":"path","required":true,"schema":{"type":"integer","title":"Draw Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/predict/":{"get":{"tags":["Prediction"],"summary":"Predict","description":"Prediction endpoint.\n\nAccess:\n    Authenticated users only.","operationId":"predict_predict__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/import/upload":{"post":{"tags":["Import"],"summary":"Upload Csv","description":"Upload and import a lottery CSV file.\n\nAdministrator only.","operationId":"upload_csv_import_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_csv_import_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/statistics/dashboard":{"get":{"tags":["Statistics"],"summary":"Get Dashboard","operationId":"get_dashboard_statistics_dashboard_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/statistics/frequency":{"get":{"tags":["Statistics"],"summary":"Get Frequency","operationId":"get_frequency_statistics_frequency_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/statistics/hot":{"get":{"tags":["Statistics"],"summary":"Get Hot Numbers","operationId":"get_hot_numbers_statistics_hot_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/statistics/cold":{"get":{"tags":["Statistics"],"summary":"Get Cold Numbers","operationId":"get_cold_numbers_statistics_cold_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/statistics/overdue":{"get":{"tags":["Statistics"],"summary":"Get Overdue Numbers","operationId":"get_overdue_numbers_statistics_overdue_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/statistics/pairs":{"get":{"tags":["Statistics"],"summary":"Get Pairs","operationId":"get_pairs_statistics_pairs_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/statistics/triplets":{"get":{"tags":["Statistics"],"summary":"Get Triplets","operationId":"get_triplets_statistics_triplets_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/statistics/even-odd":{"get":{"tags":["Statistics"],"summary":"Get Even Odd","operationId":"get_even_odd_statistics_even_odd_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/statistics/sums":{"get":{"tags":["Statistics"],"summary":"Get Sums","operationId":"get_sums_statistics_sums_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/statistics/decades":{"get":{"tags":["Statistics"],"summary":"Get Decades","operationId":"get_decades_statistics_decades_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/statistics/consecutive":{"get":{"tags":["Statistics"],"summary":"Get Consecutive","operationId":"get_consecutive_statistics_consecutive_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/ai/train":{"post":{"tags":["Artificial Intelligence"],"summary":"Train Random Forest","description":"Train all Random Forest models.\n\nAccess:\n    Administrator only.","operationId":"train_random_forest_ai_train_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_login_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_login_post"},"Body_upload_csv_import_upload_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_csv_import_upload_post"},"DashboardResponse":{"properties":{"users":{"type":"integer","title":"Users"},"draws":{"type":"integer","title":"Draws"},"models":{"type":"integer","title":"Models"},"predictions":{"type":"integer","title":"Predictions"},"accuracy":{"type":"number","title":"Accuracy"},"last_training":{"type":"string","title":"Last Training"}},"type":"object","required":["users","draws","models","predictions","accuracy","last_training"],"title":"DashboardResponse"},"DrawCreate":{"properties":{"game":{"type":"string","title":"Game","default":"Loto FDJ"},"draw_date":{"type":"string","format":"date","title":"Draw Date"},"n1":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N1"},"n2":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N2"},"n3":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N3"},"n4":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N4"},"n5":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N5"},"chance":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Chance"}},"type":"object","required":["draw_date","n1","n2","n3","n4","n5","chance"],"title":"DrawCreate"},"DrawResponse":{"properties":{"game":{"type":"string","title":"Game","default":"Loto FDJ"},"draw_date":{"type":"string","format":"date","title":"Draw Date"},"n1":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N1"},"n2":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N2"},"n3":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N3"},"n4":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N4"},"n5":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N5"},"chance":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Chance"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["draw_date","n1","n2","n3","n4","n5","chance","id","created_at"],"title":"DrawResponse"},"DrawUpdate":{"properties":{"game":{"type":"string","title":"Game","default":"Loto FDJ"},"draw_date":{"type":"string","format":"date","title":"Draw Date"},"n1":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N1"},"n2":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N2"},"n3":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N3"},"n4":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N4"},"n5":{"type":"integer","maximum":49.0,"minimum":1.0,"title":"N5"},"chance":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Chance"}},"type":"object","required":["draw_date","n1","n2","n3","n4","n5","chance"],"title":"DrawUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"Token"},"UserRegister":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","email","password"],"title":"UserRegister"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"is_active":{"type":"boolean","title":"Is Active"},"is_admin":{"type":"boolean","title":"Is Admin"}},"type":"object","required":["id","username","email","is_active","is_admin"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}